third_party/WebKit/LayoutTests/http/tests/serviceworker/unregister-then-register-new-script.html

<!DOCTYPE html>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script src="resources/test-helpers.js"></script>
<script>
var worker_url = 'resources/empty-worker.js';

async_test(function(t) {
    var scope = 'resources/scope/unregister-then-register-new-script-that-404s';
    var iframe;
    var registration;

    service_worker_unregister_and_register(t, worker_url, scope)
      .then(function(r) {
          registration = r;
          return wait_for_state(t, r.installing, 'activated');
        })
      .then(function() {
          return with_iframe(scope);
        })
      .then(function(frame) {
          iframe = frame;
          return registration.unregister();
        })
      .then(function() {
          var promise = navigator.serviceWorker.register('this-will-404',
                                                         { scope: scope });
          unload_iframe(iframe);
          return promise;
        })
      .then(
        function() {
          assert_unreached('register should reject the promise');
        },
        function() {
          return with_iframe(scope);
        })
      .then(function(frame) {
          assert_equals(frame.contentWindow.navigator.serviceWorker.controller,
                        null,
                        'document should not load with a controller');
          unload_iframe(frame);
          t.done();
        })
      .catch(unreached_rejection(t));
}, 'Registering a new script URL that 404s does not resurrect an ' +
       'unregistered registration');

async_test(function(t) {
    var scope = 'resources/scope/unregister-then-register-reject-install-worker';
    var iframe;
    var registration;

    service_worker_unregister_and_register(t, worker_url, scope)
      .then(function(r) {
          registration = r;
          return wait_for_state(t, r.installing, 'activated');
        })
      .then(function() {
          return with_iframe(scope);
        })
      .then(function(frame) {
          iframe = frame;
          return registration.unregister();
        })
      .then(function() {
          var promise = navigator.serviceWorker.register(
              'resources/reject-install-worker.js', { scope: scope });
          unload_iframe(iframe);
          return promise;
        })
      .then(function(r) {
          registration = r;
          return wait_for_state(t, r.installing, 'redundant');
        })
      .then(function() {
          return with_iframe(scope);
        })
      .then(function(frame) {
          assert_equals(frame.contentWindow.navigator.serviceWorker.controller,
                        null,
                        'document should not load with a controller');
          unload_iframe(frame);
          return registration.unregister();
        })
      .then(function() {
          t.done();
        })
      .catch(unreached_rejection(t));
  }, 'Registering a new script URL that fails to install does not resurrect ' +
       'an unregistered registration');
</script>
